+++ src/provider.ts
@@ -75,6 +75,13 @@ export class Provider implements vscode.TreeDataProvider<TreeItem> {
     }
 
     public refresh(event?: vscode.TextDocumentChangeEvent) {
+        if (!vscode.window.activeTextEditor.document) {
+            // prevent event triggering when there is no active window
+            // available or the event gets triggered/raced while the editor
+            // is switching between open windows
+            return void 0;
+        }
+
         try {
             this.getProvider().refresh(event);
         } catch (ex) {